move tic_project_ext to tools
