+++ src/studio/project.h
@@ -26,4 +26,3 @@
 
 bool tic_project_load(const char* name, const char* data, s32 size, tic_cartridge* dst);
 s32 tic_project_save(const char* name, void* data, const tic_cartridge* cart);
-bool tic_project_ext(const char* name);
